INfinite scrolling added
Now I have added infinite scrolling and random platform generation into the code.

This way when the player jumps up new platforms are made randomly and the game keeps going.

It's not complete yet I have to make it look nicer because right now the platforms are very close to each other. But i will learn that tomorrow.